+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Changelog
 
 ### Planned for v0.1.3
+ - Rename `aed_shift_count` parameter as `shift_count`. Rename the default value
+   `STARNEIG_SCHUR_DEFAULT_AED_SHIFT_COUNT` as
+   `STARNEIG_SCHUR_DEFAULT_SHIFT_COUNT`.
  - Improved performance models.
 
 ### v0.1.2